+++ CHANGELOG.md
@@ -4,8 +4,8 @@ Change Log
 ----------------
 1. Fixed off-by-1 bug in Tree widget;
 2. Allow app:navigate event listener to change window.location.hash path;
-3. Make $.overlay to support view: view object - on going; 
-4. Dynamic Themes (currently racing with onShow() measuring...) - pending;
+3. Rewired window.onerror to fire app:error event;
+4. Removed view.flyTo and added view.overlay method use use views as overlays;
 
 
 1.1.1 (2014-05-19)